pyproject.toml
[project]
name = "v-to-work"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mlx-whisper",
    "mlx-audio",
    "google-genai",
    "sounddevice",
    "pynput",
    "python-dotenv",
    "numpy",
    "scipy",
]

[project.scripts]
assistant = "main:main"
